# healy-apigateway/cmd/docs/swagger.yaml
definitions:
...
  models.PrescriptionRequest:
    description: Prescription request details
    properties:
      booking_id:
        type: integer
      doctor_id:
        type: integer
      dosage:
        type: string
      medicine:
        type: string
      notes:
        type: string
    type: object
...
  models.SetAvailability:
    description: Set availability details
    properties:
      date:
        description: e.g., "2024-06-20"
        type: string
      endtime:
        description: e.g., "17:00"
        type: string
      starttime:
        description: e.g., "09:00"
        type: string
    type: object
  response.Response:
    properties:
      data:
        description: |-
          The response data
          example: {"key": "value"}
      error:
        description: |-
          The error information, if any
          example: "Error description"
      message:
        description: |-
          The response message
          example: "Request was successful"
        type: string
    type: object
info:
  contact:
    email: support@healy-medicare.com
    name: API Support
  description: This is the API documentation for the Healy Medicare application.
  title: Healy Medicare API
  version: 1.0.0
paths:
...
  /chat:
    get:
      description: Establish WebSocket connection for chat
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: WebSocket connection for chat messages
      tags:
      - Chat
  /chat/messages:
    get:
      description: Retrieve chat messages
      parameters:
      - description: FriendID
        in: query
        name: FriendID
        required: true
        type: string
      - description: Offset
        in: query
        name: Offset
        required: true
        type: string
      - description: Limit
        in: query
        name: Limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Chat Messages
      tags:
      - Chat
...
  /doctor/patient:
    get:
      consumes:
      - application/json
      description: Retrieve booked patients for the doctor
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Booked Patients
      tags:
      - Doctor
  /doctor/patient/prescription:
    post:
      consumes:
      - application/json
      description: Create a prescription for a patient
      parameters:
      - description: Prescription Details
        in: body
        name: prescription
        required: true
        schema:
          $ref: '#/definitions/models.PrescriptionRequest'
      - description: Booking ID
        in: query
        name: booking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create Prescription
      tags:
      - Doctor
...
  /doctor/profile/availability:
    post:
      consumes:
      - application/json
      description: Set availability for the doctor
      parameters:
      - description: Availability Details
        in: body
        name: availability
        required: true
        schema:
          $ref: '#/definitions/models.SetAvailability'
      - description: Availability Details
        in: body
        name: availability
        required: true
        schema:
          $ref: '#/definitions/models.SetAvailability'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Set Doctor Availability
      tags:
      - Doctor
...
  /patient/bookdoctor:
    get:
      consumes:
      - application/json
      description: Book a doctor for a specific slot,//user browser
      parameters:
      - description: Slot ID
        in: query
        name: slot_id
        required: true
        type: integer
      - description: Patient ID
        in: query
        name: patient_id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page for payment
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Book a doctor
      tags:
      - Patient
...
  /patient/doctor/availability:
    get:
      consumes:
      - application/json
      description: Retrieve the available slots for a doctor on a given date
      parameters:
      - description: Doctor ID
        in: query
        name: doctor_id
        required: true
        type: integer
      - description: Date in YYYY-MM-DD format
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get doctor's slot availability
      tags:
      - Patient
...
  /payment_success:
    get:
      consumes:
      - application/json
      description: Verify the payment and create a calendar entry for the booking
      parameters:
      - description: Booking ID
        in: query
        name: booking_id
        required: true
        type: integer
      - description: Payment ID
        in: query
        name: payment_id
        required: true
        type: string
      - description: Razor ID
        in: query
        name: razor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify payment and create calendar entry  //redirect from book doctor
      tags:
      - Patient
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"